Title: 1까지 나누기, Time: 142ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240805/1까지 나누기/divide-by-1.py
@@ -1,8 +1,8 @@
 n = int(input())
 cnt = 0
 for i in range(1,n+1):
-    n = n/i
-    if n<=1.0:
+    n = int(n/i)
+    if n<=1:
         cnt += 1
         break
     else:
